requirements.txt
requests>=2.31.0
azure-storage-blob>=12.19.0
transformers>=4.38.2
torch>=2.2.0
pandas>=2.2.0
pyodbc>=5.1.0
SQLAlchemy>=2.0.27
python-dotenv>=1.0.0
